added userlist viewset and url path
