+++ delilahdawgapi/views/__init__.py
@@ -7,4 +7,4 @@ from .tag import TagView
 from .category import CategoryView
 from .post_reaction import PostReactionView
 from .subscription import SubscriptionView
-
+from .userlist import UserListView
