+++ delilahdawgapi/views/userlist.py
@@ -1,0 +1,65 @@
+from django.contrib.auth.models import User
+from rest_framework import status
+from rest_framework.decorators import action
+from rest_framework.viewsets import ViewSet
+from rest_framework.response import Response
+from rest_framework import serializers
+from delilahdawgapi.models import RareUser
+from django.http import HttpResponseServerError
+from delilahdawgapi.models import Subscription
+
+class UserListView(ViewSet):
+    
+    @action(methods=['post', 'delete'], detail=True)
+    def subscribe(self, request, pk=None):
+        
+        rareuser = RareUser.objects.get(user=request.auth.user)
+        
+        try: 
+            
+            rareuser = rareuser.objects.get(pk=pk)
+        except RareUser.DoesNotExist:
+            return Response(
+                {'message': 'Subscription does not exist.'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
+        if request.method == "POST":
+            try:
+                
+                rareuser.following.add(rareuser)
+                return Response({}, status=status.HTTP_201_CREATED)
+            except Exception as ex:
+                return Response({'message': ex.args[0]})
+            
+        elif request.method == "DELETE":
+            try:
+                
+                rareuser.following.remove(rareuser)
+                return Response(None, status=status.HTTP_204_NO_CONTENT)
+            except Exception as ex:
+                return Response({'message': ex.args[0]})
+    
+    def list(self, request):
+
+        
+        dogs = RareUser.objects.all()
+
+        serializer = RareUserSerializer(
+            dogs, many=True, context={'request': request}
+        )
+
+        return Response(serializer.data)
+    
+    
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = ('first_name', 'last_name', 'username')
+
+class RareUserSerializer(serializers.ModelSerializer):
+    user = UserSerializer(many=False)
+
+    class Meta:
+        model = RareUser
+        fields = ('user', 'bio', 'profile_image_url',
+        'created_on', 'active')
